+++ jason/src/rpc/backoff_delayer.rs
@@ -1,40 +1,40 @@
+//! Implementation of delayer which increases delay time by provided multiplier
+//! on every delay call.
+
 use derive_more::{Display, From};
-use js_sys::Promise;
 use tracerr::Traced;
-use wasm_bindgen_futures::JsFuture;
 
-use crate::utils::{window, JsCaused, JsDuration, JsError};
+use crate::utils::{resolve_after, JsCaused, JsDuration, JsError};
 
-/// Errors which can occur in [`ProgressiveDelayer`].
+/// Errors which can occur in [`BackoffDelayer`].
 #[derive(Debug, From, Display, JsCaused)]
 pub enum BackoffDelayerError {
     /// Error which can happen while setting JS timer.
     #[display(fmt = "{}", _0)]
     Js(JsError),
 }
-
-/// Delayer which will increase delay time in geometry progression after any
-/// `delay` calls.
+/// Delayer which increases delay time by provided multiplier on every delay
+/// call
 ///
-/// Delay time increasing will be stopped when [`ProgressiveDelayer::max_delay`]
+/// Delay time increasing will be stopped when [`BackoffDelayer::max_delay`]
 /// milliseconds of `current_delay` will be reached. First delay will be
-/// [`ProgressiveDelayer::current_delay_ms`].
+/// [`BackoffDelayer::current_delay`].
 pub struct BackoffDelayer {
-    /// Milliseconds of [`ProgressiveDelayer::delay`] call.
+    /// Delay of next [`BackoffDelayer::delay`] call.
     ///
-    /// Will be increased by [`ProgressiveDelayer::delay`] call.
+    /// Will be increased by [`BackoffDelayer::delay`] call.
     current_delay: JsDuration,
 
-    /// Max delay for which this [`ProgressiveDelayer`] may delay.
+    /// Max delay for which this [`BackoffDelayer`] may delay.
     max_delay: JsDuration,
 
-    /// The multiplier by which [`ProgressiveDelayer::current_delay`] will be
-    /// multiplied on [`ProgressiveDelayer::delay`].
+    /// The multiplier by which [`BackoffDelayer::current_delay`] will be
+    /// multiplied on [`BackoffDelayer::delay`] call.
     multiplier: f32,
 }
 
 impl BackoffDelayer {
-    /// Returns new [`ProgressiveDelayer`].
+    /// Returns new [`BackoffDelayer`].
     pub fn new(
         starting_delay_ms: JsDuration,
         multiplier: f32,
@@ -47,7 +47,7 @@ impl BackoffDelayer {
         }
     }
 
-    /// Returns next step of delay.
+    /// Returns [`JsDuration`] for a next delay.
     fn get_delay(&mut self) -> JsDuration {
         if self.is_max_delay_reached() {
             self.max_delay
@@ -58,30 +58,22 @@ impl BackoffDelayer {
         }
     }
 
-    /// Returns `true` when max delay ([`ProgressiveDelayer::max_delay_ms`]) is
+    /// Returns `true` when max delay ([`BackoffDelayer::max_delay`]) is
     /// reached.
     fn is_max_delay_reached(&self) -> bool {
         self.current_delay >= self.max_delay
     }
 
-    /// Resolves after [`ProgressiveDelayer::current_delay`] milliseconds.
+    /// Resolves after [`BackoffDelayer::current_delay`] delay.
     ///
     /// Next call of this function will delay
-    /// [`ProgressiveDelayer::current_delay_ms`] *
-    /// [`ProgressiveDelayer::multiplier`] milliseconds.
+    /// [`BackoffDelayer::current_delay`] *
+    /// [`BackoffDelayer::multiplier`] milliseconds.
     pub async fn delay(&mut self) -> Result<(), Traced<BackoffDelayerError>> {
-        let delay_ms = self.get_delay();
-        JsFuture::from(Promise::new(&mut |yes, _| {
-            window()
-                .set_timeout_with_callback_and_timeout_and_arguments_0(
-                    &yes,
-                    delay_ms.into_js_duration(),
-                )
-                .unwrap();
-        }))
-        .await
-        .map(|_| ())
-        .map_err(JsError::from)
-        .map_err(tracerr::from_and_wrap!())
+        let delay = self.get_delay();
+        resolve_after(delay)
+            .await
+            .map_err(JsError::from)
+            .map_err(tracerr::from_and_wrap!())
     }
 }